bcrypt implementation

+++ src/controllers/usercontroller.js
@@ -1,14 +1,16 @@
 const User = require('../models/user.model.js');
+const bcrypt = require('bcrypt');
+const saltRounds = 10
 
 const getUser = async (req,res)=>{
   try {
     if (req.query){
-    const users = await User.find(req.query);
+    const users = await User.find(req.query,"username name");
     res.status(200).json(users);
     return
     }
     else{
-      res.status(200).json(await User.find({}))
+      res.status(200).json(await User.find({},"username name"))
     }
   } catch (error) {
     res.status(500).json({ message: error.message });
@@ -16,9 +18,10 @@ const getUser = async (req,res)=>{
 };
 
 const createUser = async (req, res) => {
-  console.log(req.body)
   try {
-    const user = await User.create(req.body);
+    const passwordHash = await bcrypt.hash(req.body.password,saltRounds);
+    var user = await User.create({name:req.body.name,username:req.body.username,password:passwordHash});
+    user = await User.findById(user.id,"-password");
     res.status(200).json(user);
   } catch (error) {
     res.status(500).json({ message: error.message });
@@ -27,13 +30,18 @@ const createUser = async (req, res) => {
 
 const login = async(req,res)=>{
   try{
-    const user = await User.findOne({username:req.body.username,password:req.body.password});
+    const user = await User.findOne({username:req.body.username});
     if (!user){
       res.status(403).json({message:"Incorrect login credentials"});
+      return
     }
-    else{
-      res.status(200).json({message:`Logged in as ${user.name} (@${user.username})`});
+    const password = await bcrypt.compare(req.body.password,user.password)
+    if (!password){
+      res.status(403).json({message:"Incorrect login credentials"});
+      return
     }
+    res.status(200).json({success:true,message:`Successfully logged in as @${user.username}`});
+
   }
   catch(err){
     res.status(500).json({ message: error.message });
@@ -42,15 +50,18 @@ const login = async(req,res)=>{
 
 const deleteUser = async(req,res)=>{  
   try{
-    const user = await User.findOne({username:req.body.username,password:req.body.password});
+    const user = await User.findOne({username:req.body.username});
     if (!user){
       res.status(403).json({message:"Incorrect login credentials"});
+      return
     }
-    else{
-      console.log(user.id)
-      await User.findByIdAndDelete(user.id)
-      res.status(200).json({message:`Deleted ${user.name} (@${user.username})`});
+    const password = await bcrypt.compare(req.body.password,user.password)
+    if (!password){
+      res.status(403).json({message:"Incorrect login credentials"});
+      return
     }
+    await User.findByIdAndDelete(user.id)
+    res.status(200).json({message:`Deleted ${user.name} (@${user.username})`});  
   }
   catch(err){
     res.status(500).json({ message: err.message });
@@ -59,12 +70,21 @@ const deleteUser = async(req,res)=>{
 
 const editUser = async (req,res)=>{
   try{
-    const user = await User.findOne({username:req.body.username,password:req.body.password});
+    const user = await User.findOne({username:req.body.username});
     if (!user){
       res.status(403).json({message:"Incorrect login credentials"});
       return
     }
-    await User.findByIdAndUpdate(user.id,{username:req.body.newusername,name:req.body.newname,password:req.body.newpassword});
+    const password = await bcrypt.compare(req.body.password,user.password)
+    if (!password){
+      res.status(403).json({message:"Incorrect login credentials"});
+      return
+    }
+    let newData = {};
+    if(req.body.newpassword){newData.password = bcrypt.hash(req.body.newpassword,saltRounds);}
+    if (req.body.newname){newData.name = req.body.newname}
+    if (req.body.newusername){newData.username = req.body.newusername}
+    await User.findByIdAndUpdate(user.id,newData);
     res.status(200).json({message:"Successfully updated credentials"});
   }
   catch (err){
